requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
boto3==1.38.32
botocore==1.38.32
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
Django==5.2.1
django-allauth==65.8.1
django-ses==4.4.0
django-storages==1.14.6
django-tailwind==4.0.1
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
